test: move in-memory reuqests to their own file
